requirements.txt
flask
flask-sqlalchemy
flask-login
flask-caching
python-dotenv
requests
